Moved TargetType test to Destructurer test file.
